separate protocol and injector context code
