+++ packages/di/src/lib/context/injector-context.ts
@@ -1,0 +1,4 @@
+import { Injector } from '../injector.js';
+import { Context, createContext } from './protocol.js';
+
+export const INJECTOR_CTX: Context<'injector', Injector> = createContext('injector');
