+++ packages/di/src/lib/dom-injector.ts
@@ -1,4 +1,5 @@
-import { ContextRequestEvent, INJECTOR_CTX } from './context.js';
+import { ContextRequestEvent } from './context/protocol.js';
+import { INJECTOR_CTX } from './context/injector-context.js';
 import { injectables, Injector } from './injector.js';
 
 export class DOMInjector extends Injector {
